LayoutTests/editing/pasteboard/data-transfer-items-expected.txt

CONSOLE MESSAGE: line 24: Populating DataTransferItems...
CONSOLE MESSAGE: line 32: Caught exception "Error: INVALID_STATE_ERR: DOM Exception 11" as expected.
CONSOLE MESSAGE: line 36: Verifying contents of DataTransferItems...
CONSOLE MESSAGE: line 37: items.length: 2
CONSOLE MESSAGE: line 38: items[0].kind: string
CONSOLE MESSAGE: line 39: items[0].type: text/plain
CONSOLE MESSAGE: line 40: items[1].kind: string
CONSOLE MESSAGE: line 41: items[1].type: text/html
CONSOLE MESSAGE: line 47: Checking if items past the end of the collection can be indexed:
CONSOLE MESSAGE: line 48: items[2] is undefined: undefined
CONSOLE MESSAGE: line 60: Checking that a read-only DataTransferItems cannot be mutated...
CONSOLE MESSAGE: line 63: items.length: 2
CONSOLE MESSAGE: line 64: items[0].kind: string
CONSOLE MESSAGE: line 65: items[0].type: text/html
CONSOLE MESSAGE: line 68: items[1].kind: string
CONSOLE MESSAGE: line 69: items[1].type: text/plain
CONSOLE MESSAGE: line 72: items[2] is undefined: undefined
CONSOLE MESSAGE: line 89: Testing if DataTransferItems can be accessed outside an event handler...
CONSOLE MESSAGE: line 98: DataTransferItem accessed outside event handler!
CONSOLE MESSAGE: line 15: copy: items[0] value: Hello World!
CONSOLE MESSAGE: line 15: copy: items[1] value: <b>Hello World!
CONSOLE MESSAGE: line 15: paste: items[0] value: <span class="Apple-style-span" style="border-collapse: separate; color: rgb(0, 0, 0); font-family: Times; font-style: normal; font-variant: normal; font-weight: normal; letter-spacing: normal; line-height: normal; orphans: 2; text-align: -webkit-auto; text-indent: 0px; text-transform: none; white-space: normal; widows: 2; word-spacing: 0px; -webkit-border-horizontal-spacing: 0px; -webkit-border-vertical-spacing: 0px; -webkit-text-decorations-in-effect: none; -webkit-text-size-adjust: auto; -webkit-text-stroke-width: 0px; font-size: medium; ">This file tests the basic functionality and properties of DataTransferItems. This test requires DRT.</span>
CONSOLE MESSAGE: line 15: paste: items[1] value: This file tests the basic functionality and properties of DataTransferItems. This test requires DRT.
This file tests the basic functionality and properties of DataTransferItems. This test requires DRT.
